Implementation of `OtlpHttpLogExporter`
