+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_http_exporter.h
@@ -3,43 +3,22 @@
 
 #pragma once
 
-#include <chrono>
-#include <memory>
-#include <mutex>
-#include <string>
-
 // We need include exporter.h first, which will include Windows.h with NOMINMAX on Windows
 #include "opentelemetry/sdk/trace/exporter.h"
 
-#include "opentelemetry/ext/http/client/http_client.h"
+#include "opentelemetry/exporters/otlp/otlp_http_client.h"
 
 #include "opentelemetry/exporters/otlp/otlp_environment.h"
+
+#include <chrono>
+#include <memory>
+#include <string>
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace exporter
 {
 namespace otlp
 {
-// The default URL path to post metric data.
-constexpr char kDefaultMetricsPath[] = "/v1/metrics";
-// The default URL path to post metric data.
-constexpr char kDefaultLogPath[] = "/v1/logs";
-// The HTTP header "Content-Type"
-constexpr char kHttpJsonContentType[]   = "application/json";
-constexpr char kHttpBinaryContentType[] = "application/x-protobuf";
-
-enum class JsonBytesMappingKind
-{
-  kHexId,
-  kHex,
-  kBase64,
-};
-
-enum class HttpRequestContentType
-{
-  kJson,
-  kBinary,
-};
 
 /**
  * Struct to hold OTLP exporter options.
@@ -111,9 +90,6 @@ public:
   bool Shutdown(std::chrono::microseconds timeout = std::chrono::microseconds(0)) noexcept override;
 
 private:
-  // Stores if this exporter had its Shutdown() method called
-  bool is_shutdown_ = false;
-
   // For testing
   friend class OtlpHttpExporterTestPeer;
 
@@ -121,9 +97,7 @@ private:
   const OtlpHttpExporterOptions options_;
 
   // Object that stores the HTTP sessions that have been created
-  std::shared_ptr<ext::http::client::HttpClient> http_client_;
-  // Cached parsed URI
-  std::string http_uri_;
+  OtlpHttpClient http_client_;
 };
 }  // namespace otlp
 }  // namespace exporter
